Sources cleanup. Remove some warnings.
git-svn-id: https://zxtune.googlecode.com/svn/trunk@2616 e19a5227-2556-0410-816f-a35683127991

+++ src/core/plugins/players/plugin.cpp
@@ -52,7 +52,7 @@ namespace ZXTune
           callback.ProcessModule(inputData, holder);
           Parameters::IntType usedSize = 0;
           properties.GetResult()->FindValue(Module::ATTR_SIZE, usedSize);
-          return Analysis::CreateMatchedResult(usedSize);
+          return Analysis::CreateMatchedResult(static_cast<std::size_t>(usedSize));
         }
       }
       return Analysis::CreateUnmatchedResult(Decoder->GetFormat(), data);
